+++ changelog.md
@@ -16,7 +16,8 @@ Patch: Everything else → bumps patch version (1.0.0 → 1.0.1)
 
 ### Changed
 
--
+- changed generate-version.js
+- changed the AppVersion component
 
 ### Fixed